requirements.txt
# requirements.txt
python-chess==0.29.0
python-dotenv==0.20.0
openai==0.6.4
